% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{add_full_names}
\alias{add_full_names}
\title{Add full fragments names as the first column to the existed data.frame}
\usage{
add_full_names(df, addM = TRUE, addN = TRUE)
}
\arguments{
\item{df}{data.frame load with load_data function.}
}
\value{
data.frame with the first column containing full fragments names
}
\description{
Add full fragments names as the first column to the existed data.frame
}
\details{
Convenience function to use instead of get_full_names. Input data.frame has
to contain three columns names \code{FragID}, \code{M} and \code{N}
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- add_full_names(df)
}

