% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_signif}
\alias{add_signif}
\title{Add statistical significance to fragments contributions}
\usage{
add_signif(df, FUN = wilcox.test)
}
\arguments{
\item{df}{input data.frame.}

\item{FUN}{function which returns p.value within the list of named values.
By default wilcox.test is used, t.test can be used also}
}
\value{
data.frame with added two columns names pvalue and ptext:
with numerical p.value and with string representation of p.value
(*** < 0.001, ** < 0.01, * < 0.05)
}
\description{
Add statistical significance to fragments contributions
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- add_signif(df)
}
