% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clust}
\alias{clust}
\title{Clustering fragment contributions}
\usage{
clust(data, molids = NULL)
}
\arguments{
\item{data}{vector of fragment contributions.}

\item{molids}{vector of molecule IDs corresponding to fragment contributions}
}
\value{
Mclust model object or NULL if data has a single unique observation value (see details).
}
\description{
Clustering fragment contributions
}
\details{
Mclust model is a gaussian mixture model based on integrated complete-
data likelihood optimization criterion. If all values  in data are equal
 or a single value provided, then NULL is returned.
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- dplyr::filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
}
