% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clust_all}
\alias{clust_all}
\title{Build mclust models for all fragments}
\usage{
clust_all(data)
}
\arguments{
\item{data}{input data.frame}
}
\value{
list containing mclust models for fragments contained in data.frame
}
\description{
Build mclust models for all fragments
}
\details{
If all contributuons of a fragment are identical the model
will not be built.
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- dplyr::filter(df, Model == "consensus", Property == "overall")
models <- clust_all(df)
}
