% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_by_frags_count}
\alias{filter_by_frags_count}
\title{Filter out fragments which are outside of specified ranges}
\usage{
filter_by_frags_count(df, minM = 10, minN = 10, maxM = Inf, maxN = Inf)
}
\arguments{
\item{df}{input data.frame.}

\item{minM, maxM, minN, maxN}{values of minimum and maximum valus for M and N columns in the input data.frame}
}
\value{
data.frame.
}
\description{
Filter out fragments which are outside of specified ranges
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- filter_by_frags_count(df)
}
