% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_by_model_names}
\alias{filter_by_model_names}
\title{Keep only models which are specified}
\usage{
filter_by_model_names(df, keep_model_names)
}
\arguments{
\item{df}{input data.frame.}

\item{keep_model_names}{character vector of models names}
}
\value{
data.frame.
}
\description{
Keep only models which are specified
}
\details{
The input data.frame has to contain \code{Model} column.
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- filter_by_model_names(df, c("svm", "pls", "rf"))
}
