% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_by_prop_names}
\alias{filter_by_prop_names}
\title{Keep only properties which are specified}
\usage{
filter_by_prop_names(df, keep_prop_names)
}
\arguments{
\item{df}{input data.frame.}

\item{keep_prop_names}{character vector of models names}
}
\value{
data.frame.
}
\description{
Keep only properties which are specified
}
\details{
The input data.frame has to contain \code{Property} column.
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- filter_by_prop_names(df, c("svm", "pls", "rf"))
}

