% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_clust_params}
\alias{get_clust_params}
\title{Parameters of each gaussian (cluster)}
\usage{
get_clust_params(model)
}
\arguments{
\item{model}{mclust model object}
}
\value{
data.frame with mean, variance and proportion for each gaussian (cluster)
}
\description{
Parameters of each gaussian (cluster)
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- dplyr::filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
par <- get_clust_params(m)
}
