% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mol_ids}
\alias{get_mol_ids}
\title{Shows IDs of molecules for each cluster separately}
\usage{
get_mol_ids(model, uncert = 1)
}
\arguments{
\item{model}{mclust model}

\item{uncert}{numeric value, the maximum level of uncertainty for molecules to belong to either cluster. Molecules with uncertaintiy higher then this threshold   will not appear in the resulting list}
}
\value{
list of vectors containing MolIDs of molecules belonging to each cluster  (as many vectors as many clusters are in the model; elements of list are named according to clusters)
}
\description{
Shows IDs of molecules for each cluster separately
}
