% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{load_data}
\alias{load_data}
\title{Read and preprocess data about fragments contributions from a text file}
\usage{
load_data(file_name, sep = "###")
}
\arguments{
\item{file_name}{name of the input text file with fragments contributions.}

\item{sep}{separator between molecule and fragment names. Default is ###.}
}
\value{
melted data.frame with added consensus (average) contributions
}
\description{
Read and preprocess data about fragments contributions from a text file
}
\details{
An input file of an old format had '#' as a separator.
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
}

