% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_data_ext}
\alias{load_data_ext}
\title{Read and preprocess data about fragments contributions from a text file created with spci-ext Python scripts}
\usage{
load_data_ext(file_name, sep = "###")
}
\arguments{
\item{file_name}{name of the input text file with fragments contributions.}

\item{sep}{separator between molecule and fragment names. Default is ###.}
}
\value{
data.frame in long format
}
\description{
Read and preprocess data about fragments contributions from a text file created with spci-ext Python scripts
}
\details{
An input file doesn't have a header and consists of two columns with names and values
}
\examples{
df <- load_data_ext(file_name)
}
