% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_contrib}
\alias{plot_contrib}
\title{Plot fragments contributions}
\usage{
plot_contrib(df, frag_name_col = "full_name", contrib_col = "Contribution",
  plot_type = "boxplot", FUN = median, show_sign_text = NULL,
  show_sep_lines = TRUE, flip = TRUE, x_labels_angle = 60)
}
\arguments{
\item{df}{input data.frame.}

\item{frag_name_col}{name of a column with fragments names to plot}

\item{contrib_col}{name of a column with fragments contribution values}

\item{plot_type}{maybe \code{"boxplot"} or \code{"barplot"}}

\item{FUN}{aggragate function used in barplot (should return named list of outputs with numeric p.value item)}

\item{show_sign_text}{column name of text represented significance (usually number of asterisks). If \code{NULL} no text will be plot.}

\item{show_sep_lines}{boolean to control of showing grey lines to separate different fragments on a plot}

\item{flip}{boolean to control the orientation of a plot}

\item{x.labels.angle}{rorate axis x text labels. It works only if \code{flip} was set to TRUE.}
}
\value{
ggplot object which can be further modified before printing
}
\description{
Plot fragments contributions
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- reorder_data(df, "consensus", "overall")
df <- add_signif(df)
plot_contrib(df)
plot_contrib(df, plot_type = "barplot", flip = FALSE, show_sign_text = "ptext")
}
