% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_mclust_plots}
\alias{save_mclust_plots}
\title{Save plots of multiple mclust models in a grid image}
\usage{
save_mclust_plots(filename, models)
}
\arguments{
\item{filename}{file name to save plots}

\item{models}{list of mclust models}
}
\description{
Save plots of multiple mclust models in a grid image
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- dplyr::filter(df, Model == "consensus", Property == "overall")
models <- clust_all(df)
save_mclust_plots("models.png", models)
}
