% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clust}
\alias{clust}
\title{Clustering fragment contributions}
\usage{
clust(data, molids = NULL)
}
\arguments{
\item{data}{vector of fragment contributions.}

\item{molids}{vector of molecule IDs corresponding to fragment contributions}
}
\value{
Mclust model object
}
\description{
Clustering fragment contributions
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
}
