% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clust_all}
\alias{clust_all}
\title{Build mclust models for multiple fragments}
\usage{
clust_all(df, fragnames)
}
\arguments{
\item{df}{input data.frame}

\item{fragnames}{column name of the input data.frame containing fragment names
(i.e. "FragID" or "full_name")}
}
\value{
list containing mclust models
}
\description{
Build mclust models for multiple fragments
}
\details{
If all contributions of a fragment are identical the model
will not be built.
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- dplyr::filter(df, Model == "consensus", Property == "overall")
df <- add_full_names(df)
models <- clust_all(df, "full_name")
}
