% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clust_all}
\alias{clust_all}
\title{Build mclust models for multiple fragments}
\usage{
clust_all(data, fragnames, contrib_col_name = "Contribution",
  mol_col_name = "MolID")
}
\arguments{
\item{data}{input data.frame}

\item{fragnames}{column of the data.frame containing fragments names (i.e. FragID or full_name)}

\item{contrib_col_name}{name of a column with contribution values}

\item{mol_col_name}{name of a column with names (ids) of molecules}
}
\value{
list containing mclust models for fragments contained in data.frame
}
\description{
Build mclust models for multiple fragments
}
\details{
If all contributions of a fragment are identical the model for that fragment
will not be built.
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- dplyr::filter(df, Model == "consensus", Property == "overall")
df <- add_full_names(df)
models <- clust_all(df, "full_name")
}
