% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clust_all}
\alias{clust_all}
\title{Builds mclust models (gaussian mixture models) for all (if possible, see details) fragments present in input data}
\usage{
clust_all(data)
}
\arguments{
\item{data}{input  data.frame}
}
\value{
list containing mclust models for fragments contained in data.frame; each list element is named like a corresponding fragment
}
\description{
Builds mclust models (gaussian mixture models) for all (if possible, see details) fragments present in input data
}
\details{
If a data.frame  contains fragments  occuring  either in exactly one molecule  once or in exactly one molecule more than once with equal contributions,  then for such fragments mixture models will not be built (due to lack of data).Nothing will be returned for them.
}
