% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_by_prop_names}
\alias{filter_by_prop_names}
\title{Keep only properties which are specified}
\usage{
filter_by_prop_names(df, keep_prop_names = NULL, remove_prop_names = NULL)
}
\arguments{
\item{df}{input data.frame.}

\item{keep_prop_names}{character vector of models names}

\item{remove_prop_names}{character vector of models names}
}
\value{
data.frame.
}
\description{
Keep only properties which are specified
}
\details{
The input data.frame has to contain \code{Property} column.
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df1 <- filter_by_prop_names(df, "overall")
df2 <- filter_by_prop_names(df, remove_prop_names = c("overall", "dispersive"))
}

