% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mol_ids}
\alias{get_mol_ids}
\title{Get IDs of molecules for each cluster}
\usage{
get_mol_ids(model, uncertainty = 1)
}
\arguments{
\item{model}{mclust model}

\item{uncertainty}{the maximum level of uncertainty for molecules
to belong to a cluster. Molecules with uncertainty higher than this
threshold will not appear in the resulting list}
}
\value{
list of vectors containing IDs of molecules belonging to
each cluster if IDs were supplied to \link{clust}. Otherwise indices
will be returned.
}
\description{
Get IDs of molecules for each cluster
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- dplyr::filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
get_mol_ids(m, uncert = 0.2)
}
