% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_num_clust}
\alias{get_num_clust}
\title{Gives the number of clusters mclust gaussian mixture model  contains}
\usage{
get_num_clust(model)
}
\arguments{
\item{model}{mclust model object}
}
\value{
integer value: number of clusters (gaussians)
}
\description{
Gives the number of clusters mclust gaussian mixture model  contains
}
\examples{
#' ' file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
num <- get_num_clust(m)

}
