% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_num_clust}
\alias{get_num_clust}
\title{Get number of gaussians (clusters) in mclust model}
\usage{
get_num_clust(model)
}
\arguments{
\item{model}{mclust model object}
}
\value{
number of gaussians (clusters)
}
\description{
Get number of gaussians (clusters) in mclust model
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- dplyr::filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
num <- get_num_clust(m)
}
