% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_mclust}
\alias{plot_mclust}
\title{Visualize mclust model: plot the histogram of input data, dashed line: kernel density estimate and solid colored lines: gaussians corresponding to clusters found by model}
\usage{
plot_mclust(model, main = NULL, bin = 0.1)
}
\arguments{
\item{model}{Mclust gaussian mixture model}

\item{main}{Title for plot}

\item{bin}{Bin for histogram, distance between nearest breakpoints}
}
\value{
Plot with the histogram of input data & kernal density esitmate (dashed) & gaussians obtained with a model (solid)
}
\description{
Visualize mclust model: plot the histogram of input data, dashed line: kernel density estimate and solid colored lines: gaussians corresponding to clusters found by model
}
