% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_mclust}
\alias{plot_mclust}
\title{Plot mclust model}
\usage{
plot_mclust(
  model,
  binwidth = 0.1,
  title = NULL,
  xlab = "Contribution",
  ylab = "density"
)
}
\arguments{
\item{model}{mclust model object}

\item{binwidth}{width of bins on histogram}

\item{title}{plot title}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}
}
\value{
Plot with the histogram of occurence of fragment contributions,
its kernel density esitmate (dashed) and gaussians detected by a model
(color solid)
}
\description{
Plot mclust model
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
dx <- dplyr::filter(df, FragID == "OH (aliphatic)", Model == "consensus", Property == "overall")
m <- clust(dx$Contribution, dx$MolID)
plot_mclust(m)
}
