% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reorder_data}
\alias{reorder_data}
\title{Reorder fragments names represented as factor levels based on a specified metric}
\usage{
reorder_data(df, by_model, by_prop, FUN = median, frag_col_name = "full_name")
}
\arguments{
\item{df}{input data.frame.}

\item{by_model}{character name of a selected model name in column}

\item{by_prop}{character name of a selected property}

\item{FUN}{metric functions applied to order different fragments}

\item{frag_col_name}{character name of column containing desired fragments names to reorder}
}
\value{
data.frame.
}
\description{
Reorder fragments names represented as factor levels based on a specified metric
}
\details{
This function can be applied only to the melted input data.frame.
}
\examples{
file_name <- system.file("extdata", "free-wilson_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- reorder_data(df, "consensus", "overall", frag_col_name = "FragID")
}
