% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_mclust_plots}
\alias{save_mclust_plots}
\title{Save plots of multiple mclust models in a grid image}
\usage{
save_mclust_plots(filename, models)
}
\arguments{
\item{filename}{file name to save plots}

\item{models}{list of Mclust models}
}
\description{
Save plots of multiple mclust models in a grid image
}
\details{
If list contains more than 51 models, the function will return more
 than one png file, each next file will contain next (up to) 51 images. (File names will
 get suffices "_i", i = 1,2,3,...)
}
\examples{
file_name <- system.file("extdata", "BBB_frag_contributions.txt", package = "rspci")
df <- load_data(file_name)
df <- dplyr::filter(df, Model == "consensus", Property == "overall")
df <- add_full_names(df)
models <- clust_all(df, "full_name")
save_mclust_plots("models.png", models)
}
